requirements.txt
# LLM Graph Task Requirements
# Generated: 2026-01-30

# ============================================================================
# Core LLM and AI Frameworks
# ============================================================================
openai>=1.12.0
transformers>=4.36.0
torch>=2.0.0
huggingface-hub>=0.20.0
together>=1.0.0

# Optional: For faster local LLM inference
# vllm>=0.2.7

# ============================================================================
# RAG (Retrieval-Augmented Generation) Components
# ============================================================================
# langchain>=0.1.0
# langchain-chroma>=0.1.0
# langchain-community>=0.0.20
# langchain-core>=0.1.23
# langchain-huggingface>=0.0.1
llama-index>=0.9.0

# ============================================================================
# Embeddings and Similarity Search
# ============================================================================
sentence-transformers>=2.2.0
chromadb>=0.4.0

# ============================================================================
# Graph Processing and Data Analysis
# ============================================================================
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ============================================================================
# Web Scraping and HTTP
# ============================================================================
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Optional: For LeetCode scraping (may need to install from source)
# leetscrape

# ============================================================================
# Text Processing and NLP
# ============================================================================
rapidfuzz>=3.0.0
tiktoken>=0.5.0
python-slugify>=8.0.0

# ============================================================================
# PDF Processing
# ============================================================================
PyPDF2>=3.0.0
PyMuPDF>=1.23.0

# ============================================================================
# Documentation and Code Analysis
# ============================================================================
sphobjinv>=2.3.0

# ============================================================================
# Database and ORM
# ============================================================================
sqlalchemy>=2.0.0

# ============================================================================
# Utilities and Progress Tracking
# ============================================================================
tqdm>=4.65.0
python-dotenv>=1.0.0

# ============================================================================
# Optional Development Dependencies
# ============================================================================
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0

# ============================================================================
# Notes:
# ============================================================================
# 1. Some models may require additional GPU dependencies (CUDA toolkit)
# 2. For Apple Silicon (M1/M2), you may need to install torch with MPS support
# 3. vllm is optional but recommended for faster local LLM inference
# 4. leetscrape may need to be installed from GitHub if not available on PyPI
# 5. Set environment variables for API keys:
#    - OPENAI_API_KEY for OpenAI models
#    - DEEPSEEK_API_KEY for DeepSeek models
#    - HF_TOKEN for Hugging Face models
#    - TOGETHER_API_KEY for Together AI models
